Disable scroll while header loading
Instead of loading the rest of the sections after the header is loaded, disable
the scroll while the header is loading and let the other sections load. This way
the sections can be loaded while the user is waiting for the header to be done
animating. This is also better for SEO.

To avoid locking out the user from scrolling if the page loads at
'#intro-section', don't disable the viewport is not at the top of the page.