Pop up preview from icon instead of sliding down
Use DoChanges more often instead of AfterView/AfterContent since its is the
approperiate place to be changing properties that affect the view.

Make array of links and use ngFor instead of hardcoding every link. This way
simplifies the html generation since it only has to be written once.

Use an anchor-previewable for each previewable. This allows the images to load
earlier since their src would be set earlier and allows smoother animations.